Rename split helper and extract sibling helper
New name makes  role explicit: it walks outward from the selection to
the nearest <br>s and splits paragraphs there.

$splitParagraphsAtLineInnerBreaks is the second-pass helper that
explodes every <br> inside a selection into its own paragraph. It used
to live as a private method on Contents; pulling it out keeps the pair
side by side in lexical_helper.js.
